+++ src/packages/infiniteloading/demo.scss
@@ -1,35 +1,3 @@
 .demo-infinite {
   min-height: 100vh;
 }
-
-.infiniteUl,
-.infiniteUl3 {
-  height: 300px;
-  width: 100%;
-  padding: 0;
-  overflow-y: auto;
-  overflow-x: hidden;
-}
-
-.infiniteUl2 {
-  width: 100%;
-  padding: 0;
-  overflow-y: auto;
-  overflow-x: hidden;
-
-  &::-webkit-scrollbar {
-    width: 0;
-  }
-}
-
-.infiniteUl3 {
-  .infiniteLi {
-    margin: 10px 10px 0;
-    background-color: #fff;
-  }
-}
-
-.infiniteLi {
-  margin-top: 10px;
-  text-align: center;
-}
