+++ src/packages/infiniteloading/demos/h5/demo4.tsx
@@ -1,0 +1,100 @@
+import React, { useState, useEffect, CSSProperties } from 'react'
+import { Cell, InfiniteLoading } from '@nutui/nutui-react'
+import { More } from '@nutui/icons-react'
+
+const sleep = (time: number): Promise<unknown> =>
+  new Promise((resolve) => {
+    setTimeout(resolve, time)
+  })
+const InfiniteUlStyle: CSSProperties = {
+  height: '300px',
+  width: '100%',
+  padding: '0',
+  overflowY: 'auto',
+  overflowX: 'hidden',
+}
+
+const InfiniteLiStyle: CSSProperties = {
+  margin: '10px 10px 0',
+  fontSize: '14px',
+  color: 'rgba(100, 100, 100, 1)',
+  textAlign: 'center',
+  backgroundColor: '#FFF',
+}
+const Demo4 = () => {
+  const [customList, setCustomList] = useState<string[]>([])
+  const [customHasMore, setCustomHasMore] = useState(true)
+
+  useEffect(() => {
+    init()
+  }, [])
+
+  const init = () => {
+    for (let i = 0; i < 10; i++) {
+      customList.push(`${i}`)
+    }
+    setCustomList([...customList])
+  }
+
+  const customLoadMore = async () => {
+    await sleep(2000)
+    const curLen = customList.length
+    for (let i = curLen; i < curLen + 10; i++) {
+      customList.push(`${i}`)
+    }
+    if (customList.length >= 30) {
+      setCustomHasMore(false)
+    } else {
+      setCustomList([...customList])
+    }
+  }
+
+  return (
+    <>
+      <Cell>
+        <ul id="primaryScroll" style={InfiniteUlStyle}>
+          <InfiniteLoading
+            target="primaryScroll"
+            type="primary"
+            loadingText={
+              <>
+                <svg
+                  width="24"
+                  height="24"
+                  viewBox="0 0 24 24"
+                  fill="none"
+                  className="nut-infinite-bottom-tips-icons"
+                >
+                  <g clipPath="url(#clip0_878_2529)">
+                    <path
+                      d="M23.1507 10.6435C21.8958 9.29889 21.6209 7.28491 20.8022 6.60353C19.9835 5.92216 18.7824 7.00753 18.3402 7.6467C15.896 5.92216 11.9879 6.00054 11.9879 6.00054C11.9879 6.00054 8.09759 5.92216 5.6475 7.6467C5.20528 7.00753 4.01012 5.92216 3.19143 6.60353C2.37274 7.28491 2.10383 9.29889 0.848906 10.6435C-0.0892994 11.6566 -0.166985 11.952 0.215468 12.754C0.591945 13.556 2.93447 14.2193 4.58977 12.6334C4.81088 13.4595 6.67534 18 11.9938 18C17.3123 18 19.1887 13.4595 19.4039 12.6334C21.0592 14.2193 23.4077 13.556 23.7901 12.754C24.1666 11.952 24.0889 11.6566 23.1507 10.6435Z"
+                      fill="#ffffff"
+                    />
+                  </g>
+                </svg>
+                加载中
+              </>
+            }
+            loadMoreText={
+              <>
+                <More className="nut-infinite-bottom-tips-icons" />
+                没有啦~
+              </>
+            }
+            hasMore={customHasMore}
+            onLoadMore={customLoadMore}
+          >
+            {customList.map((item, index) => {
+              return (
+                <li key={index} style={InfiniteLiStyle}>
+                  {item}
+                </li>
+              )
+            })}
+          </InfiniteLoading>
+        </ul>
+      </Cell>
+    </>
+  )
+}
+export default Demo4
